+++ packages/block-editor/src/components/global-styles/typography-panel.js
@@ -20,6 +20,7 @@ import LetterSpacingControl from '../letter-spacing-control';
 import TextAlignmentControl from '../text-alignment-control';
 import TextTransformControl from '../text-transform-control';
 import TextDecorationControl from '../text-decoration-control';
+import TextShadowControl from '../text-shadow-control';
 import WritingModeControl from '../writing-mode-control';
 import { getValueFromVariable, useToolsPanelDropdownMenuProps } from './utils';
 import { setImmutably } from '../../utils/object';
@@ -39,6 +40,7 @@ export function useHasTypographyPanel( settings ) {
 	const hasTextAlign = useHasTextAlignmentControl( settings );
 	const hasTextTransform = useHasTextTransformControl( settings );
 	const hasTextDecoration = useHasTextDecorationControl( settings );
+	const hasTextShadow = useHasTextShadowControl( settings );
 	const hasWritingMode = useHasWritingModeControl( settings );
 	const hasTextColumns = useHasTextColumnsControl( settings );
 	const hasFontSize = useHasFontSizeControl( settings );
@@ -52,6 +54,7 @@ export function useHasTypographyPanel( settings ) {
 		hasTextTransform ||
 		hasFontSize ||
 		hasTextDecoration ||
+		hasTextShadow ||
 		hasWritingMode ||
 		hasTextColumns
 	);
@@ -105,6 +108,10 @@ function useHasTextAlignmentControl( settings ) {
 
 function useHasTextDecorationControl( settings ) {
 	return settings?.typography?.textDecoration;
+}
+
+function useHasTextShadowControl( settings ) {
+	return settings?.typography?.textShadow;
 }
 
 function useHasWritingModeControl( settings ) {
@@ -166,6 +173,7 @@ const DEFAULT_CONTROLS = {
 	textAlign: true,
 	textTransform: true,
 	textDecoration: true,
+	textShadow: true,
 	writingMode: true,
 	textColumns: true,
 };
@@ -376,6 +384,21 @@ export default function TypographyPanel( {
 	const hasWritingMode = () => !! value?.typography?.writingMode;
 	const resetWritingMode = () => setWritingMode( undefined );
 
+	// Text Shadow
+	const hasTextShadowControl = useHasTextShadowControl( settings );
+	const textShadow = decodeValue( inheritedValue?.typography?.textShadow );
+	const setTextShadow = ( newValue ) => {
+		onChange(
+			setImmutably(
+				value,
+				[ 'typography', 'textShadow' ],
+				newValue || undefined
+			)
+		);
+	};
+	const hasTextShadow = () => !! value?.typography?.textShadow;
+	const resetTextShadow = () => setTextShadow( undefined );
+
 	// Text Alignment
 	const hasTextAlignmentControl = useHasTextAlignmentControl( settings );
 
@@ -583,6 +606,22 @@ export default function TypographyPanel( {
 						value={ textAlign }
 						onChange={ setTextAlign }
 						size="__unstable-large"
+						__nextHasNoMarginBottom
+					/>
+				</ToolsPanelItem>
+			) }
+			{ hasTextShadowControl && (
+				<ToolsPanelItem
+					label={ __( 'Shadow' ) }
+					hasValue={ hasTextShadow }
+					onDeselect={ resetTextShadow }
+					isShownByDefault={ defaultControls.textShadow }
+					panelId={ panelId }
+				>
+					<TextShadowControl
+						value={ textShadow }
+						onChange={ setTextShadow }
+						__next40pxDefaultSize
 						__nextHasNoMarginBottom
 					/>
 				</ToolsPanelItem>