+++ packages/block-editor/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Enhancement
+
+-  `TextShadowControl`: Add support for text shadow styles in the block editor.
+
 ## 15.2.0 (2025-08-20)
 
 ## 15.1.0 (2025-08-07)